Home screen timer added.

+++ 98791b35ec5bcff0ce73edc4a9858609/Screens/Home/View/HomeView.swift
@@ -22,7 +22,7 @@ final class HomeView: BaseView {
         }
     }
     
-    var time: Int? {
+    var time: Int64? {
         didSet {
             timeLabel.text = "\(time ?? 0)s"
         }
